Refactorization between BTS implementations
