+++ pyfiles/avl_binary_tree.py
@@ -142,19 +142,3 @@ class AVLBinaryTree(BaseBST):
             return 0
 
         return tree.height
-
-    def get_depth(self) -> int:
-        return self._get_depth_recursive(tree=self.root)
-
-    def _get_depth_recursive(
-        self,
-        tree: Optional[AVLTreeNode],
-        depth: int = 0
-    ) -> int:
-        if not tree:
-            return depth
-
-        left_depth = self._get_depth_recursive(tree=tree.left, depth=depth+1)
-        right_depth = self._get_depth_recursive(tree=tree.right, depth=depth+1)
-
-        return max(depth, left_depth, right_depth)
