+++ pyfiles/base_bst.py
@@ -1,10 +1,10 @@
 from abc import ABC, abstractmethod
 from typing import Optional
-from domain import TreeNodeInterface
+from pyfiles.domain import TreeNodeInterface
 
 
 class BaseBST(ABC):
-    def __init__(self):
+    def __init__(self) -> None:
         self.root: Optional[TreeNodeInterface] = None
         self.n: int = 0
 
@@ -13,16 +13,32 @@ class BaseBST(ABC):
 
     def _search_recursive(
         self,
-        data: int, 
+        data: int,
         tree: Optional[TreeNodeInterface]
-    ) -> TreeNodeInterface:
-        if not tree:
+    ) -> Optional[TreeNodeInterface]:
+        if not tree or tree.data == data:
             return tree
-        
-        if tree.data < data:
+
+        if data < tree.data:
             return self._search_recursive(data=data, tree=tree.left)
 
         return self._search_recursive(data=data, tree=tree.right)
+
+    def get_depth(self) -> int:
+        return self._get_depth_recursive(tree=self.root)
+
+    def _get_depth_recursive(
+        self,
+        tree: Optional[TreeNodeInterface],
+        depth: int = 0
+    ) -> int:
+        if not tree:
+            return depth
+
+        left_depth = self._get_depth_recursive(tree=tree.left, depth=depth+1)
+        right_depth = self._get_depth_recursive(tree=tree.right, depth=depth+1)
+
+        return max(depth, left_depth, right_depth)
 
     @abstractmethod
     def insert(self):
